Fix JSON text area updating only when ticking checkbox
JSON text area did not update automatically when ticked. This was due to the fact that jsonAutoUpdate variable did not extend the observable property that the similarly named variable in the store did. This commit fixes that by directly using the variable (through reference) of the store and updating only that through the method in the store.