Fix calendar tiles not updating after saving JSON text area
Since calendar tiles do not directly use the same objects as in turnips list in the turnip store, the calendar tiles did not update when saving JSON text area and null prices were set to a non-null value. Add check to see if the tile needs to update its turnip.
Existing TurnipPriceStore functions refresh turnip objects in TurnipPriceDialog when turnip list in store is updated, causing the object to re-render and do the check.